+++ xmlschema/tests/test_package.py
@@ -22,7 +22,9 @@ class TestPackage(unittest.TestCase):
     def setUpClass(cls):
         cls.test_dir = os.path.dirname(__file__)
         cls.source_dir = os.path.dirname(cls.test_dir)
+        cls.package_dir = os.path.dirname(cls.source_dir)
         cls.missing_debug = re.compile(r"(\bimport\s+pdb\b|\bpdb\s*\.\s*set_trace\(\s*\)|\bprint\s*\()")
+        cls.get_version = re.compile(r"(?:\bversion|__version__)(?:\s*=\s*)(\'[^\']*\'|\"[^\"]*\")")
 
     def test_missing_debug_statements(self):
         # Exclude explicit debug statements written in the code
@@ -49,6 +51,30 @@ class TestPackage(unittest.TestCase):
 
             match = self.missing_debug.search(line)
             self.assertIsNone(match, message % (lineno, filename, match.group(0) if match else None))
+
+    def test_version(self):
+        message = "\nFound a different version at line %d or file %r: %r (maybe %r)."
+
+        files = [
+            os.path.join(self.source_dir, '__init__.py'),
+            os.path.join(self.package_dir, 'setup.py'),
+            os.path.join(self.package_dir, 'doc/conf.py'),
+        ]
+        version = filename = None
+        for line in fileinput.input(files):
+            if fileinput.isfirstline():
+                filename = fileinput.filename()
+            lineno = fileinput.filelineno()
+
+            match = self.get_version.search(line)
+            if match is not None:
+                if version is None:
+                    version = match.group(1).strip('\'\"')
+                else:
+                    self.assertTrue(
+                        version == match.group(1).strip('\'\"'),
+                        message % (lineno, filename, match.group(1).strip('\'\"'), version)
+                    )
 
 
 if __name__ == '__main__':